requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.10
requests==2.31.0
nltk==3.8.1
sumy==0.11.0
Werkzeug==2.3.6
python-dotenv==1.0.0
click==8.1.7
email-validator==2.1.0